google auth completed

+++ src/app/dashboard/dashboard.component.ts
@@ -1,5 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { Subject, takeUntil } from 'rxjs';
+import { MessageServiceService } from '../Services/message-service.service';
+import { SnackbarService } from '../Services/snackbar.service';
+import { UsersService } from '../Services/users.service';
 
 @Component({
   selector: 'app-dashboard',
@@ -8,17 +12,39 @@ import { Router } from '@angular/router';
 })
 export class DashboardComponent implements OnInit {
 
+  unsubscribe = new Subject();
   constructor(
-    private router: Router
+    private router: Router,
+    private user: UsersService,
+    private snack: SnackbarService,
+    private msg: MessageServiceService
   ) { }
 
   ngOnInit(): void {
-    
+
   }
 
-  logout = () =>{
-    localStorage.clear();
+  navigateToHome(){
     this.router.navigate(['/']);
+  }
+
+  logout = () => {
+    let auth: any = localStorage.getItem('auth');
+    if (auth) {
+      auth = JSON.parse(auth);
+      let { isGoogleAcc } = auth;
+      if (isGoogleAcc) {
+        this.user.googleLogout().pipe(takeUntil(this.unsubscribe)).subscribe(res => {
+          this.snack.openSnackBar(res.message);
+          localStorage.clear();
+          this.msg.changeMessage('logout');
+          this.router.navigate(['/']);
+        })
+      } else {
+        localStorage.clear();
+        this.router.navigate(['/']);
+      }
+    }
   }
 
 }
